flytt kafkamdc til tms-kafka-tools
